chore: clear remaining clippy warnings to -D warnings-clean
Extends the previous fmt/warning pass to land `cargo clippy --workspace
--all-targets` at zero warnings:

- Fix a vacuous range assertion in the FARGAN `synthesize_int` test: the
  function clamps to the C reference's `[-32767, 32767]`, not i16::MIN..MAX,
  so the pre-existing "i16 range" assertion was a tautology; rewrite to
  check the real contract.
- Remove `assert!(dtx_found || true)` (literal `true`) and the surrounding
  `dtx_found` bookkeeping from the DTX coverage test — the for-loop's
  `.unwrap()` already enforces the "encoding does not panic" invariant.
- Convert `assert_eq!(x, true/false)` → `assert!(x)` / `assert!(!x)` across
  encoder/decoder/api tests.
- Replace `a < lo || a > hi` with `!(lo..=hi).contains(&a)` in capi
  validation paths and one harness binary.
- `data.offset(pad_info.offset as isize)` → `data.add(pad_info.offset)`
  (offset is already `usize`).
- Needless `for i in 0..N { ... buf[i] ... }` → iterator forms
  (`iter().enumerate()`, `iter_mut()`, `copy_from_slice`) across several
  harness binaries and tests.
- Collapse nested `if let X { if cond { ... } }` into let-chain
  `if let X && cond { ... }` where fmt-stable.
- `std::iter::repeat(x).take(n)` → `std::iter::repeat_n(x, n)`.
- `x % 2 == 0` → `x.is_multiple_of(2)`.
- `if x >= N { x - N } else { 0 }` → `x.saturating_sub(N)`.
- `y > MAX { y = MAX } if y < MIN { y = MIN }` → `y.clamp(MIN, MAX)`.
- `&vec![...; N]` → `&[...; N]` where the size is a const.
- Extract long `(i32, ..., Option<i32>)` tuple into a local `type` alias
  in coverage_tests.
- Fix a doc comment so `+ 3 float-API` isn't parsed as a list continuation.

Two local `#[allow]` suppressions with in-code reasoning:
- `harness/src/bin/interframe.rs`: `#[allow(clippy::field_reassign_with_default)]`
  on a ~40-field snapshot function where inlining every assignment into the
  struct literal would triple the line count without improving readability.
- `harness/src/main.rs`: `#[allow(clippy::type_complexity)]` on a
  `Vec<(&str, Box<dyn Fn(usize) -> i16>)>` literal — a named type alias
  changes the `'static` inference for closures that capture `sample_rate`,
  so the inline form stays.

Verified: `cargo clippy --workspace --all-targets` → 0 warnings, 0 errors;
`cargo test -p ropus --lib`, `-p capi --lib`, `-p ropus-tools-core --lib`
all pass the same tests that passed before (one pre-existing persisted-seed
property-test regression from 2026-04-18 is unchanged, not introduced).

+++ harness-deep-plc/tests/state_divergence.rs
@@ -162,7 +162,7 @@ fn snapshot(c: &CRefFloatDecoder, r_dec: &RopusDecoder) -> Snapshot {
     // So the last 32 samples = [ltp_mem_length-32 .. ltp_mem_length).
     let ltp_mem_c = c.silk_ltp_mem_length();
     let tail_len = 32;
-    let offset_c = (ltp_mem_c - tail_len) as i32;
+    let offset_c = ltp_mem_c - tail_len;
     let c_out = c.silk_out_buf(offset_c, tail_len);
     // Rust mirror
     let ltp_mem_r = cs_r.ltp_mem_length;
